Flowjson/neo4j: move flowjson logic to neo4j
