+++ ivre/db/neo4j.py
@@ -32,9 +32,10 @@ import random
 import re
 import sys
 import time
-from py2neo import Graph, GraphError
+from py2neo import Graph, Node, Relationship, GraphError
 from py2neo import http
 from py2neo.database.status import TransientError
+from py2neo.types import remote
 
 http.socket_timeout = 3600
 
@@ -459,6 +460,9 @@ class Neo4jDBFlow(Neo4jDB, DBFlow):
     node_labels = ["Host", "Mac", "Wlan", "DNS", "Flow", "HTTP", "SSL", "SSH",
                    "SIP", "Modbus", "SNMP"]
 
+    # FIXME: will it work?
+    LABEL2NAME = {}
+
     def __init__(self, url):
         Neo4jDB.__init__(self, url)
         DBFlow.__init__(self)
@@ -684,6 +688,278 @@ class Neo4jDBFlow(Neo4jDB, DBFlow):
                                        accumulators=accumulators, time=time))
         return "\n".join(query)
 
+    @classmethod
+    def _cleanup_record(cls, elt):
+        for k, v in elt.iteritems():
+            if len(v) == 1 and all(x == None for x in v[0]):
+                elt[k] = None
+
+        cls.from_dbdict(cls._get_props(elt["elt"]))
+        new_meta = {}
+        for rec in elt["meta"]:
+            if rec["info"] is None and rec["link"] is None:
+                continue
+            info = rec["info"] or {}
+            info_props = cls._get_props(info)
+            link = rec["link"] or {}
+            link_tag = link.get("type", link.get("labels", [""])[0]).lower()
+            link_props = cls._get_props(link)
+            key = "%s%s" % ("_".join(label
+                                     for label in cls._get_labels(info,
+                                                                  info_props)
+                                     if label != "Intel"),
+                            "_%s" % link_tag if link_tag else "")
+            new_data = dict(("%s_%s" % (link_tag, k), v)
+                            for k, v in link_props.iteritems())
+            new_data.update(info_props)
+            new_meta.setdefault(key, []).append(new_data)
+        if new_meta:
+            elt["meta"] = new_meta
+            for reclist in new_meta.itervalues():
+                for rec in reclist:
+                    cls.from_dbdict(rec)
+        else:
+            del(elt["meta"])
+
+    def host_details(self, node_id):
+        q = """
+        MATCH (n)
+        WHERE ID(n) = {nid}
+        OPTIONAL MATCH (n)-[sr]->(infos:Intel)
+        WITH n, collect(distinct {info: infos, link: sr}) as infos
+        OPTIONAL MATCH (n)<-[:TO]-(in:Flow)<-[:SEND]-()
+        WITH n, infos,
+             COLLECT(DISTINCT [in.proto, COALESCE(in.dport, in.type)]) as in_flows
+        OPTIONAL MATCH (n)-[:SEND]->(out:Flow)-[:TO]->()
+        WITH n, infos, in_flows,
+             COLLECT(DISTINCT [out.proto, COALESCE(out.dport, out.type)]) as out_flows
+        OPTIONAL MATCH (n)-[:SEND]->(:Flow)-[:TO]->(dst:Host)
+        WITH n, infos, in_flows, out_flows,
+             COLLECT(DISTINCT dst.addr) as servers
+        OPTIONAL MATCH (n)<-[:TO]-(:Flow)<-[:SEND]-(src:Host)
+        WITH n, infos, in_flows, out_flows, servers,
+             COLLECT(DISTINCT src.addr) as clients
+        RETURN {elt: n, meta: infos,
+                in_flows: in_flows, out_flows: out_flows,
+                servers: servers, clients: clients}
+        """
+        node = dict(self.run(q, nid=node_id).evaluate())
+        self._cleanup_record(node)
+        return node
+
+    def flow_details(self, node_id):
+        q = """
+        MATCH (n)
+        WHERE ID(n) = {nid}
+        OPTIONAL MATCH (n)-[sr]->(infos:Intel)
+        WITH n, collect(distinct {info: infos, link: sr}) as infos
+        RETURN {elt: n, meta: infos}
+        """
+        node = dict(self.run(q, nid=node_id).evaluate())
+        cls._cleanup_record(node)
+        return node
+
+    @classmethod
+    def _query2cypher(cls, queries, mode="default", count=False, limit=None,
+                      skip=0):
+        limit = config.WEB_GRAPH_LIMIT if limit is None else limit
+        query = cls.query(
+            skip=skip, limit=limit,
+        )
+        for flt_type in ["node", "edge"]:
+            for flt in queries.get("%ss" % flt_type, []):
+                query.add_clause_from_filter(flt, mode=flt_type)
+
+        if mode == "default":
+            query.add_clause("""
+            WITH {elt: src, meta: [] } as src,
+                 {elt: link, meta: [] } as link,
+                 {elt: dst, meta: [] } as dst
+            """)
+            query.ret = "RETURN src, link, dst"
+            # FIXME: should be determined by the caller
+            formatter = cls._cursor2json_graph
+
+        elif mode == "talk_map":
+            query.add_clause('WITH src, dst, COUNT(link) AS t, '
+                             'COLLECT(DISTINCT LABELS(link)) AS labels, '
+                             'HEAD(COLLECT(ID(link))) AS ref')
+            query.ret = ("""
+                RETURN {elt: src, meta: []},
+                       {meta: [],
+                        elt: {
+                            data: { count: t, labels: labels },
+                            metadata: {labels: ["TALK"], id: ref}
+                        }} as F,
+                       {elt: dst, meta: []}
+            """)
+            formatter = cls._cursor2json_graph
+
+        elif mode == "flow_map":
+            query.add_clause('WITH src, dst, '
+                             'COLLECT(DISTINCT [link.proto, link.dport]) AS flows, '
+                             'HEAD(COLLECT(ID(link))) AS ref')
+            query.add_clause('WITH src, dst, flows, ref, SIZE(flows) AS t')
+            query.ret = ("""
+                RETURN {elt: src, meta: []},
+                       {meta: [],
+                        elt: {
+                            data: { count: t, flows: flows },
+                            metadata: {labels: ["MERGED_FLOWS"], id: ref}
+                        }} as F,
+                       {elt: dst, meta: []}
+            """)
+            formatter = cls._cursor2json_graph
+
+        if count:
+            query.ret = """
+                RETURN count(distinct src) as clients,
+                       count(distinct link) as flows,
+                       count(distinct dst) as servers
+            """
+            formatter = cls._cursor2count
+        else:
+            query.ret +=  " SKIP {skip} LIMIT {limit}"
+
+        sys.stderr.write("Executing query:\n%s\nWith params: %s\n" %
+                         (query.query, query.params))
+        return query, formatter
+
+    @staticmethod
+    def _flow2name(ref, labels, properties):
+        proto = properties.get("proto", "Flow")
+        attr = properties.get("dport", properties.get("type", None))
+        return "%s%s" % (proto, "/%s" % attr if attr is not None else "")
+
+    @classmethod
+    def _elt2name(cls, ref, labels, properties):
+        name = None
+        for label in labels:
+            for attr in cls.LABEL2NAME.get(label, []):
+                if isinstance(attr, str) or isinstance(attr, unicode):
+                    if attr in properties:
+                        name = properties[attr]
+                        break
+                else:
+                    # It's a function
+                    name = attr(ref, labels, properties)
+                    break
+            if name is not None:
+                break
+        if name is None:
+            name = ", ".join(labels)
+        return name
+
+    @classmethod
+    def _node2json(cls, ref, labels, properties):
+        name = cls._elt2name(ref, labels, properties)
+        return {
+            "id": ref,
+            "label": name,
+            "labels": labels,
+            "data": properties,
+            "x": random.random(),
+            "y": random.random(),
+        }
+
+    @classmethod
+    def _edge2json(cls, ref, from_ref, to_ref, labels, properties):
+        name = cls._elt2name(ref, labels, properties)
+        return {
+            "id": ref,
+            "label": name,
+            "labels": labels,
+            "data": properties,
+            "source": from_ref,
+            "target": to_ref,
+        }
+
+    @staticmethod
+    def _get_props(elt, meta=None):
+        if isinstance(elt, Node) or isinstance(elt, Relationship):
+            props = elt.properties
+        else:
+            props = elt.get("data", {})
+        if meta:
+            props["meta"] = meta
+        return props
+
+    @staticmethod
+    def _get_ref(elt, props):
+        if isinstance(elt, Node):
+            return int(remote(elt).ref.split('/', 1)[-1])
+        else:
+            return elt["metadata"]["id"]
+
+    @staticmethod
+    def _get_labels(elt, props):
+        if isinstance(elt, Node):
+            return list(elt.labels())
+        elif isinstance(elt, Relationship):
+            return [elt.type()]
+        else:
+            meta = elt["metadata"]
+            return meta["labels"] if "labels" in meta else [meta["type"]]
+
+    @classmethod
+    def _cursor2json_graph(cls, cursor):
+        """Transforms a cursor of triplets of (node, edge, node) to a graph of
+        hosts and flows. All the elements are of the form
+        {elt: <neo4j element-like>, meta: [<list of metadata>]}
+        This is an internal API that is very likely to change.
+        """
+        # Allows for a static layout
+        random.seed(0)
+        g = {"nodes": [], "edges": []}
+        done = set()
+
+        for src, edge, dst in cursor:
+            map(cls._cleanup_record, (src, edge, dst))
+            src_props = cls._get_props(src["elt"], src.get("meta"))
+            src_ref = cls._get_ref(src["elt"], src_props)
+            if src_ref not in done:
+                src_labels = cls._get_labels(src["elt"], src_props)
+                g["nodes"].append(cls._node2json(src_ref, src_labels, src_props))
+                done.add(src_ref)
+
+            dst_props = cls._get_props(dst["elt"], dst.get("meta"))
+            dst_ref = cls._get_ref(dst["elt"], dst_props)
+            if dst_ref not in done:
+                dst_labels = cls._get_labels(dst["elt"], dst_props)
+                g["nodes"].append(cls._node2json(dst_ref, dst_labels, dst_props))
+                done.add(dst_ref)
+
+            edge_props = cls._get_props(edge["elt"], edge.get("meta"))
+            edge_ref = cls._get_ref(edge["elt"], edge_props)
+            if edge_ref not in done:
+                edge_labels = cls._get_labels(edge["elt"], edge_props)
+                g["edges"].append(cls._edge2json(edge_ref, src_ref, dst_ref,
+                                                 edge_labels, edge_props))
+                done.add(edge_ref)
+        return g
+
+    @classmethod
+    def _cursor2count(cls, cursor):
+        res = cursor.next
+        # Compat py2neo < 3
+        try:
+            res = res()
+        except TypeError:
+            pass
+        return {"clients": res['clients'],
+                "flows": res['flows'],
+                "servers": res['servers']}
+
+    def query2graph(self, query, limit=None, skip=0, mode=None, count=False):
+        cypher_query, formatter = self._query2cypher(query, mode=mode,
+                                                     count=count, limit=limit,
+                                                     skip=skip)
+        t1 = time.time()
+        cypher_res = self.run(cypher_query)
+        sys.stderr.write("result in %s\n" % (time.time() - t1))
+        res = formatter(cypher_res)
+        return res
+
     def cleanup_flows(self):
         """Cleanup mistakes when predicting client/server ports"""
         self._cleanup_phase1()
@@ -800,3 +1076,8 @@ DETACH DELETE old_f
         self.db.run(q)
         if config.DEBUG:
             sys.stderr.write("Took %f secs\n" % (time.time() - tstamp))
+
+Neo4jDBFlow.LABEL2NAME.update({
+    "Host": ["addr"],
+    "Flow": [Neo4jDBFlow._flow2name],
+})
